Version 0.7! Editor render fix, added parallax, fixed move event,

+++ editor/src/de/sunnix/srpge/editor/data/GameObject.java
@@ -59,31 +59,31 @@ public class GameObject {
     private static final Color PLAYER_OBJECT_TOP_COLOR = new Color(.6f, 1f, 1f, .65f);
     private static final Color PLAYER_OBJECT_SIDE_COLOR = new Color(.4f, .75f, .75f, .65f);
 
-    public void draw(Window window, Graphics2D g, float zoom, int offsetX, int offsetY, boolean selected){
-        var TW = (int)(TILE_WIDTH * zoom);
-        var TH = (int)(TILE_HEIGHT * zoom);
-        var x = (int)(this.x * TW) + offsetX;
-        var y = (int)((this.z - this.y) * TH) + offsetY;
-        var w = (int)(this.width * TW);
-        var h = (int)(this.height * TH);
-        var d = (int)(this.width * TH);
+    public void draw(Window window, Graphics2D g, float zoom, float offsetX, float offsetY, boolean selected){
+        var TW = TILE_WIDTH * zoom;
+        var TH = TILE_HEIGHT * zoom;
+        var x = this.x * TW + offsetX;
+        var y = (this.z - this.y) * TH + offsetY;
+        var w = this.width * TW;
+        var h = this.height * TH;
+        var d = this.width * TH;
         components.forEach(comp -> comp.onDraw(window, this, g, zoom, x, y, w, h, d, selected));
         drawHitbox(g, zoom, offsetX, offsetY, selected);
     }
 
-    private void drawHitbox(Graphics2D g, float zoom, int offsetX, int offsetY, boolean selected){
-        var TW = (int)(TILE_WIDTH * zoom);
-        var TH = (int)(TILE_HEIGHT * zoom);
-        var x = (int)(this.x * TW) + offsetX;
-        var y = (int)((this.z - this.y) * TH) + offsetY;
-        var w = (int)(this.width * TW);
-        var h = (int)(this.height * TH);
-        var d = (int)(this.width * TH);
+    private void drawHitbox(Graphics2D g, float zoom, float offsetX, float offsetY, boolean selected){
+        var TW = TILE_WIDTH * zoom;
+        var TH = TILE_HEIGHT * zoom;
+        var x = this.x * TW + offsetX;
+        var y = (this.z - this.y) * TH + offsetY;
+        var w = this.width * TW;
+        var h = this.height * TH;
+        var d = this.width * TH;
         g.setColor(selected ? (enabled ? OBJECT_SIDE_COLOR_S : OBJECT_SIDE_COLOR_DISABLED_S) : getID() == 999 ? PLAYER_OBJECT_SIDE_COLOR : (enabled ? OBJECT_SIDE_COLOR : OBJECT_SIDE_COLOR_DISABLED));
-        g.fillRect(x, y, w, h);
+        g.fillRect((int)x, (int)y, (int)w, (int)h);
         y -= d;
         g.setColor(selected ? (enabled ? OBJECT_TOP_COLOR_S : OBJECT_TOP_COLOR_DISABLED_S) : getID() == 999 ? PLAYER_OBJECT_TOP_COLOR : (enabled ? OBJECT_TOP_COLOR : OBJECT_TOP_COLOR_DISABLED));
-        g.fillRect(x, y, w, d);
+        g.fillRect((int)x, (int)y, (int)w, (int)d);
     }
 
     public boolean intersects(float x, float y) {
