+++ editor/src/de/sunnix/srpge/editor/window/object/components/Component.java
@@ -56,6 +56,6 @@ public abstract class Component implements Cloneable{
         return new String[0];
     }
 
-    public void onDraw(Window window, GameObject parent, Graphics2D g, float zoom, int x, int y, int w, int h, int d, boolean selected) {}
+    public void onDraw(Window window, GameObject parent, Graphics2D g, float zoom, float x, float y, float w, float h, float d, boolean selected) {}
 
 }
